+++ CHANGELOG.md
@@ -9,6 +9,10 @@ It is expected that you keep this format strictly, since we depend on it in our 
 
 ## [Unreleased]
 
+### Fixed
+
+- `SBBIconButtonSmall`: `key` parameter is correctly handled in parent class (#473)
+
 ## [4.4.0] - 2025-12-18
 
 ### Added